refactor(codegen): extract declare_builtins into shared/builtins.rs (Round 7a)
1 fn lifted as an inherent impl on Cx into shared/builtins.rs (~2,367
lines of libcall extern declarations). Round 7a of the codegen
model-org refactor. This is the giant block of
`module.add_function("memcpy", ...)` / `module.add_function(
"lotus_bus_dispatch", ...)` style declarations that Pass A runs at
the start of every program build.

Surface bumps to pub(crate):
- const SOCKOPT_NAMES (consulted by declare_builtins to emit one
  getter per named socket constant)

codegen.rs: 25333 -> 22967 (-2366 lines).
