+++ crates/hale-codegen/src/shared/mod.rs
@@ -1,0 +1,4 @@
+//! Shared codegen helpers used across multiple model-organized modules.
+
+pub(crate) mod builtins;
+
